backend/services/email-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
resend==2.1.0
python-jose[cryptography]==3.3.0
email-validator==2.3.0

